added memory free functions to prevent memory leakage

+++ cnn/treelstm.cc
@@ -25,27 +25,27 @@ TreeLSTMBuilder::TreeLSTMBuilder(unsigned N,
   unsigned layer_input_dim = input_dim;
   for (unsigned i = 0; i < layers; ++i) {
     // i
-    Parameters* p_x2i = model->add_parameters({hidden_dim, layer_input_dim});
-    LookupParameters* p_h2i = model->add_lookup_parameters(N, {hidden_dim, hidden_dim});
-    Parameters* p_c2i = model->add_parameters({hidden_dim, hidden_dim});
-    Parameters* p_bi = model->add_parameters({hidden_dim});
+    Parameters* p_x2i = model->add_parameters({(long)hidden_dim, (long)layer_input_dim});
+    LookupParameters* p_h2i = model->add_lookup_parameters(N, {(long)hidden_dim, (long)hidden_dim});
+    Parameters* p_c2i = model->add_parameters({(long)hidden_dim, (long)hidden_dim});
+    Parameters* p_bi = model->add_parameters({(long)hidden_dim});
 
     // f
-    Parameters* p_x2f = model->add_parameters({hidden_dim, layer_input_dim});
-    LookupParameters* p_h2f = model->add_lookup_parameters(N*N, {hidden_dim, hidden_dim});
-    Parameters* p_c2f = model->add_parameters({hidden_dim, hidden_dim});
-    Parameters* p_bf = model->add_parameters({hidden_dim});
+    Parameters* p_x2f = model->add_parameters({(long)hidden_dim, (long)layer_input_dim});
+    LookupParameters* p_h2f = model->add_lookup_parameters(N*N, {(long)hidden_dim, (long)hidden_dim});
+    Parameters* p_c2f = model->add_parameters({(long)hidden_dim, (long)hidden_dim});
+    Parameters* p_bf = model->add_parameters({(long)hidden_dim});
 
     // o
-    Parameters* p_x2o = model->add_parameters({hidden_dim, layer_input_dim});
-    LookupParameters* p_h2o = model->add_lookup_parameters(N, {hidden_dim, hidden_dim});
-    Parameters* p_c2o = model->add_parameters({hidden_dim, hidden_dim});
-    Parameters* p_bo = model->add_parameters({hidden_dim});
+    Parameters* p_x2o = model->add_parameters({(long)hidden_dim, (long)layer_input_dim});
+    LookupParameters* p_h2o = model->add_lookup_parameters(N, {(long)hidden_dim, (long)hidden_dim});
+    Parameters* p_c2o = model->add_parameters({(long)hidden_dim, (long)hidden_dim});
+    Parameters* p_bo = model->add_parameters({(long)hidden_dim});
 
     // c (a.k.a. u)
-    Parameters* p_x2c = model->add_parameters({hidden_dim, layer_input_dim});
-    LookupParameters* p_h2c = model->add_lookup_parameters(N, {hidden_dim, hidden_dim});
-    Parameters* p_bc = model->add_parameters({hidden_dim});
+    Parameters* p_x2c = model->add_parameters({(long)hidden_dim, (long)layer_input_dim});
+    LookupParameters* p_h2c = model->add_lookup_parameters(N, {(long)hidden_dim, (long)hidden_dim});
+    Parameters* p_bc = model->add_parameters({(long)hidden_dim});
     layer_input_dim = hidden_dim;  // output (hidden) from 1st layer is input to next
 
     vector<Parameters*> ps = {p_x2i, p_c2i, p_bi, p_x2f, p_c2f, p_bf, p_x2o, p_c2o, p_bo, p_x2c, p_bc};
